Added documentation to Auth View class; Fixed lib install script.

+++ branches/vkapi_users/make_lib.sh
@@ -3,6 +3,20 @@
 ROOT_PATH="$(cd "$(dirname "$0")" && pwd)"
 cd "${ROOT_PATH}" || exit 1
 
+# Check script launched as root ================================================
+
+if [[ "$(whoami)" != 'root' ]]
+then
+    if [ -t 1 ]
+    then
+        sudo bash "$0" $*
+    else
+        gksudo bash "$0" $*
+    fi
+    
+    exit $?
+fi
+
 # Make include =================================================================
 
 rm -rf /usr/include/libqvkapi
@@ -12,6 +26,10 @@ cp -f include/*.h /usr/include/libqvkapi/ || exit $?
 
 # Make library =================================================================
 
-cd build/QVkApi/Debug || exit 2
+for makefile in $(find ./build -name Makefile)
+do
+    cd "${ROOT_PATH}"
+    cd "$(dirname "${makefile}")" || exit $?
 
-make install || exit $?
+    make install || exit $?
+done
